refactoring - added utils_plots.py file for plot/visualization utility functions
